Improve Godot Demo tester.

+++ godotDemoTester.ps1
@@ -1,3 +1,72 @@
+<#
+.SYNOPSIS
+    A script to test all the Godot demo projects in a folder.
+.DESCRIPTION
+    A somewhat silly script to run a bunch of Godot demo projects and capture the output.
+
+    -EditorTimeout=SECONDS (Default: 5) - How long to wait for the 2nd (non-import) editor run, with 0 meaning don't run it at all.
+    -PlayTimeout=SECONDS (Default: 10) - How long to wait for the game to run when there is no AHK script.
+
+    -WaitForUser=ALWAYS - Wait for the user to close the game before continuing, and do NOT run the AHK scripts.
+    -WaitForUser=NEVER (Default) - Run the AHK scripts when they exist, and CLOSE the game after the timeout when they don't.
+    -WaitForUser=AUTO - Run the AHK scripts when they exist, and WAIT for the user when they don't.
+
+# .PARAMETER Path
+    The path to the .
+.PARAMETER LiteralPath
+    Specifies a path to one or more locations. Unlike Path, the value of
+    LiteralPath is used exactly as it is typed. No characters are interpreted
+    as wildcards. If the path includes escape characters, enclose it in single
+    quotation marks. Single quotation marks tell Windows PowerShell not to
+    interpret any characters as escape sequences.
+.EXAMPLE
+    C:\PS>
+    <Description of example>
+.NOTES
+    Author: Keith Hill
+    Date:   June 28, 2010
+#>
+param(
+    [Parameter(Mandatory = $false)]
+    [string]$WaitForUser = "NEVER",
+
+    [Parameter(Mandatory = $false)]
+    [ValidateRange(0, [int]::MaxValue)]
+    [int]$EditorTimeout = 5,
+
+    [Parameter(Mandatory = $false)]
+    [ValidateRange(0, [int]::MaxValue)]
+    [int]$PlayTimeout = 10
+)
+
+Clear-Host
+
+function WaitForFileToBeWritable($filePath)
+{
+    $timeoutSeconds = 30
+    $endTime = (Get-Date).AddSeconds($timeoutSeconds)
+
+    while ((Get-Date) -lt $endTime)
+    {
+        try
+        {
+            $fileStream = [System.IO.File]::OpenWrite($filePath)
+            $fileStream.Close()
+            break
+        }
+        catch
+        {
+            Write-Host "File $filePath is not writable, retrying in 1 second..." -ForegroundColor Red -BackgroundColor black
+            Start-Sleep -Seconds 1
+        }
+    }
+
+    if ((Get-Date) -ge $endTime)
+    {
+        Write-Host "Timed out waiting for file $filePath to become writable." -ForegroundColor Red -BackgroundColor black
+    }
+}
+
 function RemoveJunkLinesFrom($inputFilePath)
 {
     # Define the list of strings
@@ -16,8 +85,10 @@ function RemoveJunkLinesFrom($inputFilePath)
         $containsString = $false
 
         # Check each string in the list
-        foreach ($string in $stringList) {
-            if ($line -match $string) {
+        foreach ($string in $stringList)
+        {
+            if ($line -match $string)
+            {
                 Write-Host "Ignoring line: $line"
                 # If the line contains the string, set the flag to true and break the loop
                 $containsString = $true
@@ -26,12 +97,14 @@ function RemoveJunkLinesFrom($inputFilePath)
         }
 
         # If the line does not contain any of the strings, write it to the output file
-        if (-not $containsString) {
+        if (-not $containsString)
+        {
             $line | Out-File -FilePath $outputFilePath -Append
         }
     }
 
     # Overwrite input file with new content
+    WaitForFileToBeWritable $inputFilePath
     Get-Content -Path $outputFilePath | Out-File -FilePath $inputFilePath
 }
 
@@ -46,7 +119,8 @@ function TestGodotDemoProject($projectFolder, $godotExecutablePath, $outputFile,
 
     # Remove the .godot folder, because that is how users will expect to initially load them.
     # Otherwise this may leave traces of my previous loading of the project.
-    if (Test-Path '.godot') {
+    if (Test-Path '.godot')
+    {
         Remove-Item -Recurse -Force '.godot'
     }
 
@@ -64,39 +138,42 @@ function TestGodotDemoProject($projectFolder, $godotExecutablePath, $outputFile,
     Start-Sleep -Seconds 1 # Give it a moment to release all file locks.
 
     # Now run it once in the editor to look for errors that only happen then.
-    $arguments = "-e"
-    $process = Start-Process $godotExecutablePath -ArgumentList $arguments -PassThru -RedirectStandardOutput $stdOutLog -RedirectStandardError $stdErrLog
-    # Wait a few seconds for it to start up, and settle, then cose it.
-    $countDown = 5
-    while ($countDown -gt 0)
+    if ($EditorTimeout -gt 0)
     {
-        Start-Sleep -Seconds 1
-        $countDown--
-        if ($process.HasExited)
+        $arguments = "-e"
+        $process = Start-Process $godotExecutablePath -ArgumentList $arguments -PassThru -RedirectStandardOutput $stdOutLog -RedirectStandardError $stdErrLog
+        # Wait a few seconds for it to start up, and settle, then cose it.
+        $countDown = $EditorTimeout
+        while ($countDown -gt 0)
         {
-            break
+            Start-Sleep -Seconds 1
+            $countDown--
+            if ($process.HasExited)
+            {
+                break
+            }
         }
-    }
-    # Check if the process is still running
-    if (!$process.HasExited)
-    {
-        # Stop the process
-        Stop-Process -Id $process.Id
-    }
-    Start-Sleep -Seconds 1 # Give it a moment to release all file locks.
+        # Check if the process is still running
+        if (!$process.HasExited)
+        {
+            # Stop the process
+            Stop-Process -Id $process.Id
+        }
+        Start-Sleep -Seconds 1 # Give it a moment to release all file locks.
 
-    # Check for and capture any errors.
-    RemoveJunkLinesFrom $stdOutLog
-    $stdOutLineCount = (Get-Content -Path $stdOutLog | Measure-Object -Line).Lines
-    if ($stdOutLineCount -gt 2)
-    {
-        Get-Content -Path $stdOutLog | Select-Object -Skip 2 | Out-File -FilePath $tempData -Append
-    }
-    RemoveJunkLinesFrom $stdErrLog
-    $stdErrLineCount = (Get-Content -Path $stdErrLog | Measure-Object -Line).Lines
-    if ($stdErrLineCount -gt 0)
-    {
-        Get-Content -Path $stdErrLog | Out-File -FilePath $tempData -Append
+        # Check for and capture any errors.
+        RemoveJunkLinesFrom $stdOutLog
+        $stdOutLineCount = (Get-Content -Path $stdOutLog | Measure-Object -Line).Lines
+        if ($stdOutLineCount -gt 2)
+        {
+            Get-Content -Path $stdOutLog | Select-Object -Skip 2 | Out-File -FilePath $tempData -Append
+        }
+        RemoveJunkLinesFrom $stdErrLog
+        $stdErrLineCount = (Get-Content -Path $stdErrLog | Measure-Object -Line).Lines
+        if ($stdErrLineCount -gt 0)
+        {
+            Get-Content -Path $stdErrLog | Out-File -FilePath $tempData -Append
+        }
     }
 
     # And run it one last time starting the game up instead of using the editor.
@@ -106,12 +183,19 @@ function TestGodotDemoProject($projectFolder, $godotExecutablePath, $outputFile,
     $scriptDirectory = $PSScriptRoot
     $testingScriptFolder = $projectFolder.ToString().Replace($godotDemoProjectFolder, "$scriptDirectory\godotGameTesting\")
     $testScript = "$testingScriptFolder\test.ahk"
-    if (Test-Path $testScript) {
+    if ((Test-Path $testScript) -and ($WaitForUser -ne "ALWAYS"))
+    {
         # Run the AutoHotkey script to test the game.
         Start-Process AutoHotkey64.exe $testScript
-    } else {
+    }
+    elseif ($WaitForUser -ne "NEVER")
+    {
+        Write-Host "Waiting for user to close the game..." -ForegroundColor Blue -BackgroundColor black
+    }
+    else
+    {
         # Otherwise just let the game run for a few seconds and then kill it.
-        $countDown = 10
+        $countDown = $PlayTimeout
         while ($countDown -gt 0)
         {
             Start-Sleep -Seconds 1
@@ -132,7 +216,7 @@ function TestGodotDemoProject($projectFolder, $godotExecutablePath, $outputFile,
 
     while (!$process.HasExited)
     {
-        # Wait for the import to finish.
+        # Wait for the process to be shut down one way or another.
         Start-Sleep -Seconds 1
     }
     Start-Sleep -Seconds 1 # Give it a moment to release all file locks.
@@ -171,11 +255,11 @@ function TestGodotDemoProjects($demoProjectFolders, $resultFile)
         $outputFile2 = "d:\temp\godotDemoTesterMaster.log"
 
         #        $godotVersionName = "4.2 Stable"
-#        $godotExecutablePath = "C:\Users\chris\OneDrive\allWindows\GodotEngines\Godot_v4.2.1-stable_win64.exe\Godot_v4.2-stable_win64.exe"
-#        $outputFile = $outputFile1
-#        TestGodotDemoProject -projectFolder $projectFolder -godotExecutablePath $godotExecutablePath -outputFile $outputFile -godotVersionName $godotVersionName
-#        "**********************************************************************" | Out-File -FilePath $resultFile -Append
-#        Get-Content -Path $outputFile | Out-File -FilePath $resultFile -Append
+        #        $godotExecutablePath = "C:\Users\chris\OneDrive\allWindows\GodotEngines\Godot_v4.2.1-stable_win64.exe\Godot_v4.2-stable_win64.exe"
+        #        $outputFile = $outputFile1
+        #        TestGodotDemoProject -projectFolder $projectFolder -godotExecutablePath $godotExecutablePath -outputFile $outputFile -godotVersionName $godotVersionName
+        #        "**********************************************************************" | Out-File -FilePath $resultFile -Append
+        #        Get-Content -Path $outputFile | Out-File -FilePath $resultFile -Append
 
         $godotVersionName = "4.3 Master"
         $godotExecutablePath = "C:\Users\chris\CLionProjects\godot\bin\godot.windows.editor.x86_64.exe"
@@ -188,10 +272,8 @@ function TestGodotDemoProjects($demoProjectFolders, $resultFile)
 
 $godotDemoProjectFolder = "D:\godot-demo-projects-clean\"
 
-Clear-Host
-
 Push-Location $godotDemoProjectFolder
-git cherry-pick --abort
+#git cherry-pick --abort
 gh repo sync chrisl8/godot-demo-projects
 git checkout master
 git reset --hard HEAD
@@ -224,16 +306,19 @@ Pop-Location
 
 $skipMonoDemos = $true
 
-if ($skipMonoDemos) {
+if ($skipMonoDemos)
+{
     #Use Mob.tscn instead of prject.godot to test with only two demos.
     #$demoProjectFolders = Get-ChildItem -Path $godotDemoProjectFolder -Filter "Mob.tscn" -Recurse -ErrorAction SilentlyContinue | Where-Object { $_.Directory.FullName -notmatch '\\mono\\' } | Select-Object -Property Directory -Unique
     $demoProjectFolders = Get-ChildItem -Path $godotDemoProjectFolder -Filter "project.godot" -Recurse -ErrorAction SilentlyContinue | Where-Object { $_.Directory.FullName -notmatch '\\mono\\' } | Where-Object { $_.Directory.FullName -notmatch '\\xr\\' } | Select-Object -Property Directory -Unique
-} else
+}
+else
 {
     $demoProjectFolders = Get-ChildItem -Path $godotDemoProjectFolder -Filter "project.godot" -Recurse -ErrorAction SilentlyContinue | Select-Object -Property Directory -Unique
 }
 
-if (-not (Test-Path -Path "D:\temp\")) {
+if (-not (Test-Path -Path "D:\temp\"))
+{
     New-Item -ItemType Directory -Path "D:\temp\"
 }
 
@@ -242,8 +327,22 @@ New-Item -ItemType File -Path $resultFile -Force > $null
 
 TestGodotDemoProjects -demoProjectFolders $demoProjectFolders -resultFile $resultFile
 
-if ($skipMonoDemos) {
+if ($skipMonoDemos)
+{
     Write-Host "Mono and XR Demos were skipped." -ForegroundColor Yellow -BackgroundColor black
 }
 
 Get-Content -Path $resultFile
+
+# TODO: Options:
+# --compare=version - Run each game twice, once with the stable version and once with the master version, allowing user to compare.
+# --include-xr - Run the XR games.
+# --include-mono - Run the Mono games.
+
+# The report output file should be opened automatically when the script finishes and not dumped to the console.
+
+# Improve ability to run this without setting stuff up ahead of time in special directories.
+
+# Use more test projects:
+# https://github.com/gdquest-demos
+# https://github.com/godotengine/godot-benchmarks
